+++ lib/i18next.dart
@@ -1,5 +1,5 @@
 library i18next;
 
 export 'src/i18next.dart';
-export 'src/interpolation_options.dart';
+export 'src/options.dart';
 export 'src/utils.dart';